backend/requirements.txt
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
supabase>=2.0.0
